[CHORE]#7: 오류를 수정해요
